WIP on properly restoring exercise controller when finished.
This was a problem because current() could not be called when plans were
finished so the screen's labels were set to placeholder values. This
commit adds an explicit done flag to plans which should mean that the
finished case is OK but we're getting asserts because we're trying to call
current() before starting the plan. Might be able to fix this by checking
isStarted but I think it'd be better to improve how plans track state.