Fixed HLM2 so that exercises at a rep range are (mostly) decoupled from the higher range.
This is what we want to do because it should be possible to advance at a lower rep
range even if you're stuck at a higher rep range. In addition it's now a lot easier
to see what you're doing at each rep range and to change what you're doing at a rep
range.